requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
plotly==5.17.0

# Financial Data and Trading
yfinance==0.2.28
ccxt==4.1.77
ta==0.10.2
backtrader==1.9.78.123
vectorbt==0.26.0
empyrical==0.5.5

# Web Framework and Dashboard
dash==2.16.1
dash-bootstrap-components==1.5.0

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Networking and HTTP
requests==2.31.0
urllib3==2.0.7

# Scheduling and Automation
schedule==1.2.0

# Data Processing
openpyxl==3.1.2

# Security and Cryptography
cryptography==41.0.7

# Real-time Data and WebSockets
websockets==12.0
python-telegram-bot==20.7

# Database and Storage
sqlite3

# Testing Framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Advanced ML Libraries
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
optuna==3.4.0
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
shap==0.44.0
mlflow==2.8.1
wandb==0.16.1
transformers==4.36.2
datasets==2.16.1
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
alpha-vantage==2.3.1
newsapi-python==0.2.6
tweepy==4.14.0
praw==7.7.1 